Tested Needed For Viewing Stats
I think I fixed viewing the stats now, but it still needs to be tested. I also modified the code so it reduces the number of redundant lines, as well as modifying main.cpp so it does not get overwritten by the .ioc. I also put the code in Game_Start into Playing_Mode, since it that code happens just right after, and Game_Start state is now available to use. I suggest using that instead of the the turn state, or just switching them out.